#393 feat: Create traits.enums, add TraitsNumberTypes
