+++ src/shared/constants/enums/enums.ts
@@ -179,24 +179,6 @@ export enum WearableTypes {
     Pet = 6,
     Background = 7
 }
-export enum TraitsDefinitionsIds {
-    HPCapacity = 0,
-    APCapacity = 1,
-    DefensePower = 2,
-    ActionSpeed = 3,
-    Regen = 4,
-    Ethereality = 5,
-    MeleePower = 6,
-    RangedPower = 7,
-    TBC = 8
-}
-
-export enum TraitsEffectsTypes {
-    Name = 0,
-    Range = 1,
-    Increase = 2,
-    Decrease = 3
-}
 
 export enum TypenameType {
     ERC1155Listing = 'ERC1155Listing'
