+++ src/shared/constants/enums/filter-component.enums.ts
