+++ src/shared/constants/enums/traits.enums.ts
@@ -1,0 +1,27 @@
+export enum TraitsNumberTypes {
+    NRG = 0,
+    AGG = 1,
+    SPK = 2,
+    BRN = 3,
+    EYS = 4,
+    EYC = 5
+}
+
+export enum TraitsDefinitionsIds {
+    HPCapacity = 0,
+    APCapacity = 1,
+    DefensePower = 2,
+    ActionSpeed = 3,
+    Regen = 4,
+    Ethereality = 5,
+    MeleePower = 6,
+    RangedPower = 7,
+    TBC = 8
+}
+
+export enum TraitsEffectsTypes {
+    Name = 0,
+    Range = 1,
+    Increase = 2,
+    Decrease = 3
+}
